fix: add esm and cjs paths to browser map for overriden files
See https://github.com/evanw/esbuild/issues/1443 for discussion.

In brief if you have a project that is built by ipjs and you have
file A that imports file B, esbuild (maybe others?) resolves file B
using it's full path within the module (e.g. including `esm` or `cjs`),
which it then uses as a key to look up overrides in the `browser` map
in your `package.json`.

If there is an override it'll fail to be found because ipjs doesn't
add paths prefixed with `esm` or `cjs` to the browser map where
overrides have been delcared in the `exports` map, so the change here
is to add files with their full path prefix to the `browser` map.