fix: require regular version in stubs
When we stub files for environments that do not handle export maps
very well, we `require` a transpiled CJS implementation and re-export
it in a stub file that should be immediately overriden by the `browser`
or `exports` fields in `package.json`.

At the moment we use the value from the `browser` overrides to work out
which file we should `require` the stub. The change here is to `require`
the transpiled regular CJS version in the stub instead of the browser
override.

Turns out when Jest has `testEnvironment: 'node'` set it resolves `require`s
to the stubs, ignoreing `exports` which loads browser-specific code into
node with predictably incendiary results.